ai_server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
pillow==10.4.0
numpy==1.26.4
opencv-python-headless==4.10.0.84
